Move the PaymentInfo stuff from Footer.tsx into its own file PaymentInfo.tsx
